huwei_0806 发布系统代码

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/IIbzMyTerritoryService.java
@@ -11,6 +11,7 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -29,9 +30,9 @@ public interface IIbzMyTerritoryService extends IService<IbzMyTerritory>{
     void createBatch(List<IbzMyTerritory> list) ;
     boolean update(IbzMyTerritory et) ;
     void updateBatch(List<IbzMyTerritory> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    IbzMyTerritory get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    IbzMyTerritory get(Long key) ;
     IbzMyTerritory getDraft(IbzMyTerritory et) ;
     boolean checkKey(IbzMyTerritory et) ;
     IbzMyTerritory mobMenuCount(IbzMyTerritory et) ;
