+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/impl/ProjectStatsServiceImpl.java
@@ -79,19 +79,19 @@ public class ProjectStatsServiceImpl extends ServiceImpl<ProjectStatsMapper, Pro
 
     @Override
     @Transactional
-    public boolean remove(BigInteger key) {
+    public boolean remove(Long key) {
         boolean result=removeById(key);
         return result ;
     }
 
     @Override
-    public void removeBatch(Collection<BigInteger> idList) {
+    public void removeBatch(Collection<Long> idList) {
         removeByIds(idList);
     }
 
     @Override
     @Transactional
-    public ProjectStats get(BigInteger key) {
+    public ProjectStats get(Long key) {
         ProjectStats et = getById(key);
         if(et==null){
             et=new ProjectStats();
