+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IBlockService.java
@@ -11,6 +11,7 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -29,9 +30,9 @@ public interface IBlockService extends IService<Block>{
     void createBatch(List<Block> list) ;
     boolean update(Block et) ;
     void updateBatch(List<Block> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    Block get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    Block get(Long key) ;
     Block getDraft(Block et) ;
     boolean checkKey(Block et) ;
     boolean save(Block et) ;
