+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/ITestTaskService.java
@@ -11,6 +11,7 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -29,9 +30,9 @@ public interface ITestTaskService extends IService<TestTask>{
     void createBatch(List<TestTask> list) ;
     boolean update(TestTask et) ;
     void updateBatch(List<TestTask> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    TestTask get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    TestTask get(Long key) ;
     TestTask getDraft(TestTask et) ;
     TestTask activate(TestTask et) ;
     TestTask block(TestTask et) ;
@@ -43,12 +44,12 @@ public interface ITestTaskService extends IService<TestTask>{
     TestTask start(TestTask et) ;
     TestTask unlinkCase(TestTask et) ;
     Page<TestTask> searchDefault(TestTaskSearchContext context) ;
-    List<TestTask> selectByBuild(BigInteger id) ;
-    void removeByBuild(BigInteger id) ;
-    List<TestTask> selectByProduct(BigInteger id) ;
-    void removeByProduct(BigInteger id) ;
-    List<TestTask> selectByProject(BigInteger id) ;
-    void removeByProject(BigInteger id) ;
+    List<TestTask> selectByBuild(Long id) ;
+    void removeByBuild(Long id) ;
+    List<TestTask> selectByProduct(Long id) ;
+    void removeByProduct(Long id) ;
+    List<TestTask> selectByProject(Long id) ;
+    void removeByProject(Long id) ;
     /**
      *自定义查询SQL
      * @param sql  select * from table where id =#{et.param}
