+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/ITodoService.java
@@ -11,6 +11,7 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -29,9 +30,9 @@ public interface ITodoService extends IService<Todo>{
     void createBatch(List<Todo> list) ;
     boolean update(Todo et) ;
     void updateBatch(List<Todo> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    Todo get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    Todo get(Long key) ;
     Todo getDraft(Todo et) ;
     Todo activate(Todo et) ;
     Todo assignTo(Todo et) ;
