+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/impl/TodoServiceImpl.java
@@ -108,7 +108,7 @@ public class TodoServiceImpl extends ServiceImpl<TodoMapper, Todo> implements IT
     }
     @Override
     @Transactional
-    public Todo get(BigInteger key) {
+    public Todo get(Long key) {
         Todo tempET=new Todo();
         tempET.set("id",key);
         Todo et = getById(key);
