+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/impl/UserServiceImpl.java
@@ -79,19 +79,19 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements IU
 
     @Override
     @Transactional
-    public boolean remove(BigInteger key) {
+    public boolean remove(Long key) {
         boolean result=removeById(key);
         return result ;
     }
 
     @Override
-    public void removeBatch(Collection<BigInteger> idList) {
+    public void removeBatch(Collection<Long> idList) {
         removeByIds(idList);
     }
 
     @Override
     @Transactional
-    public User get(BigInteger key) {
+    public User get(Long key) {
         User et = getById(key);
         if(et==null){
             et=new User();
