+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/impl/UserTplServiceImpl.java
@@ -101,7 +101,7 @@ public class UserTplServiceImpl extends ServiceImpl<UserTplMapper, UserTpl> impl
     }
     @Override
     @Transactional
-    public UserTpl get(BigInteger key) {
+    public UserTpl get(Long key) {
         UserTpl et = getById(key);
         if(et==null){
             et=new UserTpl();
